@@ -1,104 +1,116 @@
 {
-  "name": "@coreui/coreui-free-angular-admin-template",
-  "version": "2.5.2",
-  "description": "CoreUI Free Angular 2+ Admin Template",
-  "author": {
-    "name": "Łukasz Holeczek",
-    "url": "http://holeczek.pl",
-    "github": "https://github.com/mrholek",
-    "twitter": "https://twitter.com/lukaszholeczek"
-  },
-  "contributors": [
-    {
-      "name": "Andrzej Kopański",
-      "url": "https://github.com/xidedix"
-    }
-  ],
-  "homepage": "https://coreui.io/angular",
-  "copyright": "Copyright 2019 creativeLabs Łukasz Holeczek",
-  "license": "MIT",
+  "name": "egret",
+  "version": "11.0.0",
+  "license": "https://themeforest.net/licenses/terms/regular",
   "scripts": {
+    "postinstall": "ngcc",
     "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build --prod --aot --buildOptimizer --commonChunk --vendorChunk --optimization --progress",
+    "start": "node --max_old_space_size=8192 node_modules/@angular/cli/bin/ng serve -o",
+    "serveProd": "ng serve --environment=prod",
+    "build": "ng build --prod=true --aot=true --extractCss=true --optimization=true --environment=prod",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "format:fix": "pretty-quick --staged",
+    "format:check": "prettier --config ./.prettierrc --list-different \"src/{app,environments,assets}/**/*{.ts,.js,.json,.css,.scss}\"",
+    "precommit": "run-s format:fix lint",
+    "lint": "tslint --fix -c ./tslint.json 'src/**/*{.ts}'"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.2",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.0.2",
-    "@angular/compiler": "^8.0.2",
-    "@angular/core": "^8.0.2",
-    "@angular/forms": "^8.0.2",
-    "@angular/http": "^8.0.0-beta.10",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.0.2",
-    "@angular/platform-browser-dynamic": "^8.0.2",
-    "@angular/router": "^8.0.2",
-    "@coreui/angular": "^2.5.2",
-    "@coreui/coreui": "^2.1.12",
-    "@coreui/coreui-plugin-chartjs-custom-tooltips": "^1.3.1",
-    "@coreui/icons": "0.3.0",
-    "@devexpress/analytics-core": "19.2-next",
-    "alertifyjs": "^1.12.0",
-    "angular-datatables": "^8.0.0",
-    "bootstrap": "^4.3.1",
-    "chart.js": "^2.8.0",
-    "core-js": "^2.6.9",
-    "datatables.net": "^1.10.20",
-    "datatables.net-dt": "^1.10.20",
-    "devexpress-reporting-angular": "19.2-next",
+    "@agm/core": "1.1.0",
+    "@angular/animations": "11.0.4",
+    "@angular/cdk": "11.0.2",
+    "@angular/common": "11.0.4",
+    "@angular/compiler": "11.0.4",
+    "@angular/core": "11.0.4",
+    "@angular/flex-layout": "9.0.0-beta.29",
+    "@angular/forms": "11.0.4",
+    "@angular/localize": "^11.0.4",
+    "@angular/material": "11.0.2",
+    "@angular/material-moment-adapter": "11.0.2",
+    "@angular/platform-browser": "11.0.4",
+    "@angular/platform-browser-dynamic": "11.0.4",
+    "@angular/router": "11.0.4",
+    "@ngu/carousel": "3.0.2",
+    "@ngx-translate/core": "13.0.0",
+    "@ngx-translate/http-loader": "6.0.0",
+    "@swimlane/ngx-datatable": "17.0.0",
+    "alertifyjs": "^1.13.1",
+    "angular-calendar": "0.28.16",
+    "angular-datatables": "^11.1.2",
+    "angular-in-memory-web-api": "0.11.0",
+    "angular-star-rating": "4.0.0-beta.3",
     "devextreme": "19.2-next",
-    "flag-icon-css": "^3.3.0",
-    "font-awesome": "^4.7.0",
-    "hammerjs": "^2.0.8",
+    "@devexpress/analytics-core": "19.2-next",
+    "devexpress-reporting-angular": "19.2-next",
+    "apexcharts": "^3.19.3",
+    "chart.js": "2.9.3",
+    "classlist.js": "1.1.20150312",
+    "core-js": "2.4.1",
+    "css-star-rating": "git+https://git@github.com/uilibrary/css-star-rating.git",
+    "date-fns": "2.14.0",
+    "echarts": "4.8.0",
+    "flag-icon-css": "3.5.0",
+    "highlight.js": "10.1.1",
+    "hopscotch": "0.3.1",
+    "install": "0.12.2",
     "jquery": "^3.4.1",
-    "jspdf": "^1.5.2",
     "jwt-decode": "^2.2.0",
-    "mat-timepicker": "^2.0.2",
     "moment": "^2.24.0",
-    "mutationobserver-shim": "^0.3.3",
-    "ng2-charts": "^2.3.0",
-    "ng2-file-upload": "^1.4.0",
-    "ngx-bootstrap": "^5.0.0",
-    "ngx-perfect-scrollbar": "^8.0.0",
-    "rxjs": "^6.5.2",
-    "simple-line-icons": "^2.4.1",
-    "ts-helpers": "^1.1.2",
-    "tsickle": "^0.35.0",
-    "tslib": "^1.10.0",
-    "web-animations-js": "^2.3.1",
-    "xlsx": "^0.14.1",
-    "zone.js": "^0.9.1"
+    "ng-apexcharts": "^1.5.0",
+    "ng2-charts": "2.3.2",
+    "ng2-dragula": "2.1.1",
+    "ng2-file-upload": "1.4.0",
+    "ngx-bootstrap": "^6.2.0",
+    "ngx-color-picker": "^10.1.0",
+    "ngx-custom-validators": "^8.0.0",
+    "ngx-echarts": "5.0.0",
+    "ngx-material-timepicker": "^5.5.3",
+    "ngx-pagination": "5.0.0",
+    "ngx-perfect-scrollbar": "^9.0.0",
+    "ngx-quill": "12.0.1",
+    "npm": "6.13.7",
+    "perfect-scrollbar": "1.5.0",
+    "quill": "1.3.7",
+    "rxjs": "6.5.4",
+    "rxjs-compat": "6.1.0",
+    "tinycolor2": "^1.4.1",
+    "tslib": "^2.0.0",
+    "web-animations-js": "^2.3.2",
+    "xlsx": "^0.16.9",
+    "zone.js": "0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.800.3",
-    "@angular/cli": "^8.0.3",
-    "@angular/compiler-cli": "^8.0.2",
-    "@angular/language-service": "^8.0.2",
+    "@angular-devkit/build-angular": "~0.1100.4",
+    "@angular-devkit/core": "11.0.4",
+    "@angular/cli": "11.0.4",
+    "@angular/compiler-cli": "11.0.4",
+    "@angular/language-service": "11.0.4",
     "@types/datatables.net": "^1.10.18",
-    "@types/jasmine": "^3.3.13",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/jquery": "^3.3.31",
-    "@types/node": "^12.0.8",
-    "codelyzer": "^5.1.0",
-    "jasmine-core": "^3.4.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^4.1.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "^5.4.2",
-    "ts-node": "^8.3.0",
-    "tslint": "^5.17.0",
-    "typescript": "~3.4.5"
-  },
-  "engines": {
-    "node": ">= 12",
-    "npm": ">= 6"
+    "@types/hopscotch": "0.2.30",
+    "@types/jasmine": "~3.6.0",
+    "@types/jquery": "^3.3.33",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "copy-webpack-plugin": "4.3.0",
+    "enhanced-resolve": "3.3.0",
+    "husky": "^4.3.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.1.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "node-sass": "^4.14.1",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.1.2",
+    "pretty-quick": "^3.0.2",
+    "protractor": "~7.0.0",
+    "ts-node": "2.0.0",
+    "tslint": "~6.1.0",
+    "tslint-angular": "^3.0.3",
+    "typescript": "4.0.5"
   }
 }
