@@ -27,18 +27,68 @@ namespace EssPortal.Concrete
             _configuration = configuration;
         }
         
-
-        public async Task<List<Notification>> GetNotification(int UserID)
+        public async Task<List<dynamic>> GetNotification(int UserID, String NotType)
         {
             using (SqlConnection con = new SqlConnection(_configuration.GetConnectionString("DatabaseConnection")))
             {
                 var paramater = new DynamicParameters(); 
                 paramater.Add("@UserID", UserID);
+                paramater.Add("@NotType", NotType);
 
-                IEnumerable<Notification> notifications = await con.QueryAsync<Notification>("[ESS].LoadNotificationInfo", paramater, null, 0, commandType: CommandType.StoredProcedure);
-                return notifications.ToList();
-
-            }
+                return con.Query("[ess].LoadNotificationInfo", paramater, null, true, 0, commandType: System.Data.CommandType.StoredProcedure).ToList();
+             }
         }
+             
+        public async Task<bool> SaveNotification(int pNotDecisionID,bool pNotStatus,string pNotType, int pUserID ,int pNotActionID, int? pCreatedUserID)
+        {
+            try
+            {
+                Notification notification = new Notification
+                {
+                    NotDecisionID = pNotDecisionID,
+                    NotStatus = pNotStatus,
+                    NotType = pNotType,
+                    CreatedUserID = pCreatedUserID,
+                    NotActionID = pNotActionID,
+                    UserID = pUserID
+                };
+                await _context.Notifications.AddAsync(notification);
+                var result = await _context.SaveChangesAsync();
+                if (result > 0)
+                {
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
+            }
+            catch (Exception)
+            { 
+            }
+            return false;
+        }
+
+        public async Task<bool> HideNotification(int Id)
+        {
+            try
+            {
+                var obj = _context.Notifications.Where(p => p.ID == Id).FirstOrDefault();
+                if (obj != null)
+                {
+                    obj.NotStatus = true;
+                     _context.Notifications.Update(obj);
+                    await _context.SaveChangesAsync();
+                }               
+                return true;
+            }
+            catch (Exception)
+            {
+                return false;
+                throw;
+            }
+            return false;
+        }
+
     }
 }
