@@ -1,6 +1,7 @@
 ﻿
 
 using EssPortal.Models;
+using EssPortal.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -10,7 +11,9 @@ namespace EssPortal.Interfaces
 {
     public interface IEmployeeRepository : IGenericRepository<Employee>
     {
-      
+        Task<List<Employee>> GetByManager(string EmployeeID, int UserID); 
+        Task<string> CheckEmployeeIsHR(string EmployeeID, int UserID);
+        Task<List<EmployeeUserHRViewModel>> GetEmployeeUserHR();
     }
 
 }
