@@ -6,13 +6,17 @@ using System.Text;
 
 namespace EssPortal.Models
 {
-  
-    
+    [Table("Notification", Schema = "ess")]
     public class Notification
     {
-        public int NotiCount { get; set; }
-        public string MessageAR { get; set; }
-        public string MessageEN { get; set; }
+        [Key]
+        public int ID { get; set; }
+        public int UserID { get; set; } 
+        public string NotType { get; set; } 
+        public int NotDecisionID { get; set; }
+        public int NotActionID { get; set; }        
+        public Boolean NotStatus { get; set; }
+        public int? CreatedUserID { get; set; }
 
     }
 }
