@@ -1,25 +1,41 @@
-import { UploadFilesComponent } from './filesmanager/SingleFile/UploadFiles.component';
-import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { EmployeeComponent } from './employee/employee.component';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { FilesmanagerModule } from './filesmanager/filesmanager.module';
 
+// SERVICES
+import { ThemeService } from './services/theme.service';
+import { NavigationService } from "./services/navigation.service";
+import { RoutePartsService } from './services/route-parts.service';
+import { AuthGuard } from './guards/auth.guard';
+import { UserRoleGuard } from './guards/user-role.guard';
+import { AppConfirmService } from './services/app-confirm/app-confirm.service';
+import { AppLoaderService } from './services/app-loader/app-loader.service';
+
+import { SharedComponentsModule } from './components/shared-components.module';
+import { SharedPipesModule } from './pipes/shared-pipes.module';
+import { SharedDirectivesModule } from './directives/shared-directives.module'; 
+ 
 
 @NgModule({
   imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    FilesmanagerModule
+  CommonModule,
+    SharedComponentsModule,
+    SharedPipesModule,
+    SharedDirectivesModule    , 
+     
   ],
-  declarations: [
-    EmployeeComponent,
+  providers: [
+    ThemeService,
+    NavigationService,
+    RoutePartsService,
+    AuthGuard,
+    UserRoleGuard,
+    AppConfirmService,
+    AppLoaderService
   ],
-  providers: [],
   exports: [
-    EmployeeComponent,
-    FilesmanagerModule,
+    SharedComponentsModule,
+    SharedPipesModule,
+    SharedDirectivesModule, 
   ]
 })
 export class SharedModule { }
