@@ -1,20 +1,53 @@
-import { Component, OnInit } from '@angular/core';
-import { Router, NavigationEnd } from '@angular/router';
+import { Component, OnInit, AfterViewInit, Renderer2 } from '@angular/core';
+import { Title } from '@angular/platform-browser';
+import { Router, NavigationEnd, ActivatedRoute, ActivatedRouteSnapshot } from '@angular/router';
+
+import { RoutePartsService } from './shared/services/route-parts.service';
+
+import { filter } from 'rxjs/operators';
+import { UILibIconService } from './shared/services/ui-lib-icon.service';
+import { AlertifyService } from './core/services';
 
 @Component({
-  // tslint:disable-next-line
-  selector: 'body',
-  template: '<router-outlet></router-outlet>'
+  selector: 'app-root',
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.css']
 })
-export class AppComponent implements OnInit {
-  constructor(private router: Router) { }
+export class AppComponent implements OnInit, AfterViewInit {
+  appTitle = 'Egret';
+  pageTitle = '';
+
+  constructor(
+    public title: Title,
+    private router: Router,
+    private activeRoute: ActivatedRoute,
+    private routePartsService: RoutePartsService, 
+    private iconService: UILibIconService
+  ) {
+ 
+    iconService.init()
+  }
 
   ngOnInit() {
-    this.router.events.subscribe((evt) => {
-      if (!(evt instanceof NavigationEnd)) {
-        return;
+    this.changePageTitle();
+  }
+
+  ngAfterViewInit() {
+  }
+
+  changePageTitle() {
+    this.router.events.pipe(filter(event => event instanceof NavigationEnd)).subscribe((routeChange) => {
+      const routeParts = this.routePartsService.generateRouteParts(this.activeRoute.snapshot);
+      if (!routeParts.length) {
+        return this.title.setTitle(this.appTitle);
       }
-      window.scrollTo(0, 0);
+      // Extract title from parts;
+      this.pageTitle = routeParts
+                      .reverse()
+                      .map((part) => part.title )
+                      .reduce((partA, partI) => {return `${partA} > ${partI}`});
+      this.pageTitle += ` | ${this.appTitle}`;
+      this.title.setTitle(this.pageTitle);
     });
   }
 }
