@@ -1,133 +1,177 @@
-import { TestComponent } from './views/test/test.component';
-
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-// Import Containers
-import { DefaultLayoutComponent } from './containers';
-import { P404Component } from './views/error/404.component';
-import { P500Component } from './views/error/500.component';
-import { LoginComponent } from './views/login/login.component';
-import { AuthGuard } from './auth/AuthGuard';
-import { UserLogoutComponent } from './views/login/app.UserLogout.Component';
-import { AccessDeniedComponent } from './views/error/access-denied/access-denied.component';
+import { Routes } from '@angular/router';
 import { AdminAuthGuardService } from './auth/AdminAuthGuardService';
+import { AdminLayoutComponent } from './shared/components/layouts/admin-layout/admin-layout.component';
+import { AuthLayoutComponent } from './shared/components/layouts/auth-layout/auth-layout.component';
+import { AuthGuard } from './shared/guards/auth.guard';
 
-export const routes: Routes = [
+export const rootRouterConfig: Routes = [
   {
     path: '',
     redirectTo: 'dashboard',
-    pathMatch: 'full',
-  },
+    pathMatch: 'full'
+  }, 
   {
-    path: '404',
-    component: P404Component,
-    data: {
-      title: 'Page 404'
-    }
+    path: 'home',
+    loadChildren: () => import('./views/home/home.module').then(m => m.HomeModule),
+    data: { title: 'Choose A Demo' }
   },
-  {
-    path: '500',
-    component: P500Component,
-    canActivate: [AuthGuard],
-    data: {
-      title: 'Page 500'
-    }
-  },
-  {
-    path: 'access-denied',
-    component: AccessDeniedComponent,
-    data: {
-      title: 'access-denied'
-    }
-  },
-  {
-    path: 'login',
-    component: LoginComponent,
-    data: {
-      title: 'Login Page'
-    }
-  },
-  { path: 'UserLogout', component: UserLogoutComponent },
-  { path: 'test', component: TestComponent },
-  // {
-  //   path: 'register',
-  //   component: RegisterComponent,
-  //   canActivate: [AuthGuard],
-  //   data: {
-  //     title: 'Register Page'
-  //   }
-  // },
   {
     path: '',
-    component: DefaultLayoutComponent,
-    data: {
-      title: 'Home'
-    },
-    runGuardsAndResolvers: 'always',
-    canActivateChild: [AuthGuard],
+    component: AuthLayoutComponent,
     children: [
       {
-        path: 'Orders',
-        loadChildren: () => import('./views/Orders/Orders.module').then(m => m.OrdersModule),
-        canLoad: [AuthGuard]
-      },
+        path: 'sessions',
+        loadChildren: () => import('./views/sessions/sessions.module').then(m => m.SessionsModule),
+        data: { title: 'Session'}
+      }
+    ]
+  },
+  {
+    path: '',
+    component: AdminLayoutComponent,
+    canActivate: [AuthGuard],
+    children: [
       {
-        path: 'Purchases',
-        loadChildren: () => import('./views/purchases/purchases.module').then(m => m.PurchasesModule),
-        canLoad: [AuthGuard]
+        canActivateChild: [AuthGuard],
+        path: 'dashboard',
+        loadChildren: () => import('./views/dashboard/dashboard.module').then(m => m.DashboardModule),
+        data: { title: 'Dashboard', breadcrumb: 'DASHBOARD'}
       },
-      {
-        path: 'PersonalInfo',
-        loadChildren: () => import('./views/PersonalInfo/personalInfo.module').then(m => m.PersonalInfoModule),
-
-      },
-      {
-        canActivate: [AdminAuthGuardService],
+      { 
+        canActivateChild: [AuthGuard],
         path: 'AdminArea',
         loadChildren: () => import('./views/AdminArea/AdminArea.module').then(m => m.AdminAreaModule),
         canLoad: [AuthGuard]
       },
       {
-        path: 'base',
-        loadChildren: () => import('./views/base/base.module').then(m => m.BaseModule),
+       canActivateChild: [AuthGuard],
+        path: 'Orders',
+        loadChildren: () => import('./views/Orders/Orders.module').then(m => m.OrdersModule),
+        data: { title: 'Orders', breadcrumb: 'ORDERS'},
         canLoad: [AuthGuard]
       },
       {
-        path: 'buttons',
-        loadChildren: () => import('./views/buttons/buttons.module').then(m => m.ButtonsModule)
+        canActivateChild: [AuthGuard],
+         path: 'Evaluation',
+         loadChildren: () => import('./views/evaluation/evaluation.module').then(m => m.EvaluationModule),
+         data: { title: 'Evaluation', breadcrumb: 'EVALUATION'},
+         canLoad: [AuthGuard]
+       },
+      {
+        path: 'material',
+        loadChildren: () => import('./views/material-example-view/material-example-view.module').then(m => m.MaterialExampleViewModule),
+        data: { title: 'Material', breadcrumb: 'MATERIAL'}
+      },
+      {
+        path: 'dialogs',
+        loadChildren: () => import('./views/app-dialogs/app-dialogs.module').then(m => m.AppDialogsModule),
+        data: { title: 'Dialogs', breadcrumb: 'DIALOGS'}
+      },
+      {
+        path: 'profile',
+        loadChildren: () => import('./views/profile/profile.module').then(m => m.ProfileModule),
+        data: { title: 'Profile', breadcrumb: 'PROFILE'}
+      },
+      {
+        path: 'others',
+        loadChildren: () => import('./views/others/others.module').then(m => m.OthersModule),
+        data: { title: 'Others', breadcrumb: 'OTHERS'}
+      },
+      {
+        path: 'tables',
+        loadChildren: () => import('./views/tables/tables.module').then(m => m.TablesModule),
+        data: { title: 'Tables', breadcrumb: 'TABLES'}
+      },
+      {
+        path: 'tour',
+        loadChildren: () => import('./views/app-tour/app-tour.module').then(m => m.AppTourModule),
+        data: { title: 'Tour', breadcrumb: 'TOUR'}
+      },
+      {
+        path: 'forms',
+        loadChildren: () => import('./views/forms/forms.module').then(m => m.AppFormsModule),
+        data: { title: 'Forms', breadcrumb: 'FORMS'}
+      },
+      {
+        path: 'chart',
+        loadChildren: () => import('./views/chart-example-view/chart-example-view.module').then(m => m.ChartExampleViewModule),
+        data: { title: 'Charts', breadcrumb: 'CHARTS'}
       },
       {
         path: 'charts',
-        loadChildren: () => import('./views/chartjs/chartjs.module').then(m => m.ChartJSModule)
+        loadChildren: () => import('./views/charts/charts.module').then(m => m.AppChartsModule),
+        data: { title: 'Charts', breadcrumb: 'CHARTS'}
       },
       {
-        path: 'dashboard',
-        loadChildren: () => import('./views/dashboard/dashboard.module').then(m => m.DashboardModule)
+        path: 'map',
+        loadChildren: () => import('./views/map/map.module').then(m => m.AppMapModule),
+        data: { title: 'Map', breadcrumb: 'MAP'}
+      },
+      {
+        path: 'dragndrop',
+        loadChildren: () => import('./views/dragndrop/dragndrop.module').then(m => m.DragndropModule),
+        data: { title: 'Drag and Drop', breadcrumb: 'DND'}
+      },
+      {
+        path: 'inbox',
+        loadChildren: () => import('./views/app-inbox/app-inbox.module').then(m => m.AppInboxModule),
+        data: { title: 'Inbox', breadcrumb: 'INBOX'}
+      },
+      {
+        path: 'calendar',
+        loadChildren: () => import('./views/app-calendar/app-calendar.module').then(m => m.AppCalendarModule),
+        data: { title: 'Calendar', breadcrumb: 'CALENDAR'}
+      },
+      {
+        path: 'chat',
+        loadChildren: () => import('./views/app-chats/app-chats.module').then(m => m.AppChatsModule),
+        data: { title: 'Chat', breadcrumb: 'CHAT'}
+      },
+      {
+        path: 'cruds',
+        loadChildren: () => import('./views/cruds/cruds.module').then(m => m.CrudsModule),
+        data: { title: 'CRUDs', breadcrumb: 'CRUDs'}
+      },
+      {
+        path: 'shop',
+        loadChildren: () => import('./views/shop/shop.module').then(m => m.ShopModule),
+        data: { title: 'Shop', breadcrumb: 'SHOP'}
+      },
+      {
+        path: 'search',
+        loadChildren: () => import('./views/search-view/search-view.module').then(m => m.SearchViewModule)
+      },
+      {
+        path: 'invoice',
+        loadChildren: () => import('./views/invoice/invoice.module').then(m => m.InvoiceModule)
+      },
+      {
+        path: 'todo',
+        loadChildren: () => import('./views/todo/todo.module').then(m => m.TodoModule)
+      },
+      {
+        path: 'orders',
+        loadChildren: () => import('./views/order/order.module').then(m => m.OrderModule),
+        data: { title: 'Orders', breadcrumb: 'Orders'}
+      },
+      {
+        path: 'page-layouts',
+        loadChildren: () => import('./views/page-layouts/page-layouts.module').then(m => m.PageLayoutsModule)
+      },
+      {
+        path: 'utilities',
+        loadChildren: () => import('./views/utilities/utilities.module').then(m => m.UtilitiesModule)
       },
       {
         path: 'icons',
-        loadChildren: () => import('./views/icons/icons.module').then(m => m.IconsModule)
-      },
-      {
-        path: 'notifications',
-        loadChildren: () => import('./views/notifications/notifications.module').then(m => m.NotificationsModule)
-      },
-      {
-        path: 'theme',
-        loadChildren: () => import('./views/theme/theme.module').then(m => m.ThemeModule)
-      },
-      {
-        path: 'widgets',
-        loadChildren: () => import('./views/widgets/widgets.module').then(m => m.WidgetsModule)
+        loadChildren: () => import('./views/mat-icons/mat-icons.module').then(m => m.MatIconsModule),
+        data: { title: 'Icons', breadcrumb: 'Icons'}
       }
     ]
   },
-  { path: '**', component: P404Component }
+  {
+    path: '**',
+    redirectTo: 'sessions/404'
+  }
 ];
 
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
-})
-export class AppRoutingModule { }
