@@ -1,7 +1,9 @@
+import { HelpersService } from './services/helpers.service';
 import { NotificationService } from './services/Notification.service';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { TranslateModule } from '@ngx-translate/core';
 
 import {
   ApiService,
@@ -15,7 +17,8 @@ import { JwtInterceptor } from './_helpers/jwt.Interceptor';
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule,
+   
   ],
   providers: [
     { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
@@ -23,8 +26,10 @@ import { JwtInterceptor } from './_helpers/jwt.Interceptor';
     AlertifyService,
     AuthService,
     EssPortalService,
+    HelpersService,
     JwtService,
-    NotificationService
+    NotificationService,
+    TranslateModule
   ],
   declarations: []
 })
