@@ -1,18 +1,32 @@
-import { ChangePasswordComponent } from './change-password/change-password.component';
+
 import { RegisterComponent } from './register/register.component';
 import { AdminAreaRoutesModule } from './AdminArea.routing';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { DataTablesModule } from 'angular-datatables';
-import { AssignRoleComponent } from './Role/Role.component';
-import { AllAssignRoleComponent } from './AllRole/AllRole.component';
+import { AssignRoleComponent } from './Role/Role.component'; 
 import { SettingComponent } from './Settings/Setting.component';
 import { ShowUsersComponent } from './AllUser/ShowUser.component';
-import { EditUserRegistrationComponent } from './EditUserRegistration/EditUserRegistration.component';
+import { PermissionOperationComponent } from './permission-operation/permission-operation.component';
+
 import { MatStepperModule } from '@angular/material/stepper';
 import { MatInputModule } from '@angular/material/input';
-import { AdminService } from './Services/admin.service';
+import { AdminService } from './Services/admin.service'; 
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatSortModule } from '@angular/material/sort';
+import { MatTableDataSource, MatTableModule } from '@angular/material/table';
+import { NgxDatatableModule } from '@swimlane/ngx-datatable';
+import { TranslateModule } from '@ngx-translate/core';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatCardModule } from '@angular/material/card';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import {MatDividerModule} from '@angular/material/divider';
+import {MatSelectModule} from '@angular/material/select';
+import { EditUserRegistrationComponent } from './EditUserRegistration/EditUserRegistration.component';
+
 
 @NgModule({
   imports: [
@@ -22,7 +36,19 @@ import { AdminService } from './Services/admin.service';
     ReactiveFormsModule,
     DataTablesModule,
     MatStepperModule,
-    MatInputModule
+    MatPaginatorModule,
+    MatSortModule,
+    MatTableModule,
+    NgxDatatableModule,
+    MatInputModule,
+    MatCheckboxModule,
+    MatButtonModule,
+    MatCardModule,
+    MatIconModule,
+    MatDividerModule,
+    FlexLayoutModule,
+    MatSelectModule,
+    TranslateModule
   ],
   declarations: [
     ShowUsersComponent,
@@ -30,9 +56,8 @@ import { AdminService } from './Services/admin.service';
     RegisterComponent,
     EditUserRegistrationComponent,
     AssignRoleComponent,
-    AllAssignRoleComponent,
-    ChangePasswordComponent,
-
+    PermissionOperationComponent,
+    
   ],
   providers: [AdminService]
 
