@@ -1,6 +1,6 @@
-import { ChangePasswordComponent } from './change-password/change-password.component';
-
-import { AllAssignRoleComponent } from './AllRole/AllRole.component';
+import { PermissionOperationComponent } from './permission-operation/permission-operation.component';
+import { AdminAuthGuardService } from './../../auth/AdminAuthGuardService';
+import { ChangePasswordComponent } from '../profile/change-password/change-password.component';  
 import { RegisterComponent } from './register/register.component';
 import { Routes, RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
@@ -9,12 +9,12 @@ import { ShowUsersComponent } from './AllUser/ShowUser.component';
 import { SettingComponent } from './Settings/Setting.component';
 import { EditUserRegistrationComponent } from './EditUserRegistration/EditUserRegistration.component';
 
-
 const routes: Routes = [
   {
     path: '',
     //component: RegisterComponent,
     // canActivate: [AuthGuard],
+    canActivate: [AdminAuthGuardService],
     children: [
       { path: 'User/Add', component: RegisterComponent },
       { path: 'User/Edit/:UserId', component: EditUserRegistrationComponent },
@@ -22,13 +22,7 @@ const routes: Routes = [
     ]
   },
   {
-    path: 'ChangePassword',
-    component: ChangePasswordComponent,
-    data: {
-      title: 'Setting Portal'
-    }
-  },
-  {
+    canActivate: [AdminAuthGuardService],
     path: 'Setting',
     component: SettingComponent,
     data: {
@@ -39,11 +33,20 @@ const routes: Routes = [
     path: 'Assign',
     // component: ShowUsersComponent,
     // canActivate: [AuthGuard],
+    canActivate: [AdminAuthGuardService],
     children: [
       { path: 'Role', component: AssignRoleComponent },
-      { path: 'AllRole', component: AllAssignRoleComponent }
+      // { path: 'AllRole', component: AllAssignRoleComponent }
     ]
-  },
+  }, 
+  {
+    canActivate: [AdminAuthGuardService],
+    path: 'Permission',
+    component: PermissionOperationComponent,
+    data: {
+      title: 'Permission'
+    }
+  }
 ];
 
 
