@@ -1,22 +1,73 @@
-import { TestModule } from './../test/test.module';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatChipsModule } from '@angular/material/chips';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatTableModule } from '@angular/material/table';
+import { MatTabsModule } from '@angular/material/tabs';
+import { RouterModule } from '@angular/router';
+import { FlexLayoutModule } from '@angular/flex-layout';
 import { ChartsModule } from 'ng2-charts';
-import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
-import { ButtonsModule } from 'ngx-bootstrap/buttons';
+import { NgxEchartsModule } from 'ngx-echarts';
+import * as echarts from 'echarts';
 
-import { DashboardComponent } from './dashboard.component';
-import { DashboardRoutingModule } from './dashboard-routing.module';
+import { NgApexchartsModule } from 'ng-apexcharts';
+import { NgxDatatableModule } from '@swimlane/ngx-datatable';
+import { SharedPipesModule } from 'app/shared/pipes/shared-pipes.module';
 
+import { DashboardRoutes } from './dashboard.routing';
+import { AnalyticsComponent } from './analytics/analytics.component';
+import { DashboardDarkComponent } from './dashboard-dark/dashboard-dark.component';
+import { CryptocurrencyComponent } from './cryptocurrency/cryptocurrency.component';
+import { DefaultDashboardComponent } from './default-dashboard/default-dashboard.component';
+import { LearningManagementComponent } from './learning-management/learning-management.component';
+import { AnalyticsAltComponent } from './analytics-alt/analytics-alt.component';
+import { TranslateModule } from '@ngx-translate/core'; 
+import { TranslateService } from '@ngx-translate/core';
+import { DashBoardService } from './Services/DashBoard.service';
 @NgModule({
   imports: [
-    FormsModule,
-    DashboardRoutingModule,
+    CommonModule,
+    MatIconModule,
+    MatCardModule,
+    MatMenuModule,
+    MatProgressBarModule,
+    MatExpansionModule,
+    MatButtonModule,
+    MatChipsModule,
+    MatListModule,
+    MatTabsModule,
+    MatTableModule,
+    MatGridListModule,
+    FlexLayoutModule,
+    TranslateModule,
     ChartsModule,
-    BsDropdownModule,
-    ButtonsModule.forRoot(),
-    TestModule
+    NgxEchartsModule.forRoot({
+      echarts
+    }),
+    NgApexchartsModule,
+    NgxDatatableModule,
+    SharedPipesModule,
+    RouterModule.forChild(DashboardRoutes)
   ],
-  declarations: [DashboardComponent]
+  declarations: [
+    AnalyticsComponent,
+    DashboardDarkComponent,
+    CryptocurrencyComponent,
+    DefaultDashboardComponent,
+    LearningManagementComponent,
+    AnalyticsAltComponent],
+  exports: [],
+  providers: [
+      TranslateService,
+      DashBoardService]
 })
-export class DashboardModule { }
+export class DashboardModule {
+
+}
