@@ -1,4 +1,3 @@
-import 'hammerjs';
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
@@ -9,8 +8,5 @@ if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic().bootstrapModule(AppModule, {
-  useJit: true,
-  preserveWhitespaces: true
-})
+platformBrowserDynamic().bootstrapModule(AppModule)
   .catch(err => console.log(err));
